changed styling for nav login dropdown, refactored App.js
