+++ src/components/Header.css
@@ -1,10 +1,15 @@
 /* for tablet*/
-@media only screen and (max-width: 600px) {
+@media only screen and (max-width: 599px) {
   #nav-user-info {
     display: none;
   }
   #nav-auth-button {
     display: none;
+  }
+
+  #nav-dropdown-menu {
+    padding: 10px;
+    margin: 10px -7px 5px 0;
   }
 }
 
@@ -14,6 +19,11 @@
   }
   #nav-auth-icon-dropdown {
     display: none;
+  }
+
+  #nav-dropdown-menu {
+    padding: 10px;
+    margin: 18px -7px 5px 0;
   }
 }
 
@@ -130,4 +140,5 @@
 
 #nav-dropdown-menu {
   width: 342px;
+  padding: 10px;
 }
