+++ src/components/HomePage.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import './HomePage.css'
+import Login from './Login'
+
+export default function HomePage() {
+  return (
+    <div>
+      <Login />
+    </div>
+  )
+}
